backend/pyproject.toml
[project]
name = "agent-ui-backend"
version = "0.1.0"
description = "FastAPI backend with MCP SDK for AI Agent UI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "pyyaml>=6.0.0",
    "httpx>=0.28.0",
    "msal>=1.31.0",
    "mcp[sse]>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
